refactor(federated): #301 keep engine.go under the 500-line limit
The doc block added for the DSN quoting fix pushed engine.go to 515
lines. coding-standard.md makes 500 a hard limit, and the only approved
exception in this PR is internal/httpapi/server.go (#220) — introducing
a standards violation in the one file this PR touched for a security
reason is not a trade worth making.

Move DuckDBPostgresConnStringFromPool and its documentation to
internal/federated/connstring.go rather than trimming the comment. The
comment is load-bearing now that quoting, escaping and redaction depend
on one another, and a small single-purpose file matches internal/pgdsn.

engine.go drops to 471, below its 493-line base; connstring.go is 50.
